[CELEBORN-21][REFACTOR] Extract revive related logical from LifecycleManager (#1024)
* [CELEBORN-21][REFACTOR] Extract revive related logical from LifecycleManager
